# Conflicts:
#	BE/BE/build.gradle
#	BE/BE/src/main/java/com/linkrap/BE/entity/Scrap.java
#	BE/BE/src/main/java/com/linkrap/BE/repository/ScrapRepository.java
#	BE/BE/src/main/resources/application.properties

+++ BE/BE/src/main/java/com/linkrap/BE/repository/ScrapRepository.java
@@ -3,4 +3,9 @@ package com.linkrap.BE.repository;
 import com.linkrap.BE.entity.Scrap;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ScrapRepository extends CrudRepository<Scrap, Long> {}
+import java.util.ArrayList;
+import java.util.Optional;
+
+public interface ScrapRepository extends CrudRepository<Scrap, Integer> {
+
+}
